# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ecommerce API",
    "description": "Ecommerce API",
    "contact": {
      "name": "Taha Shabaan",
      "email": "tahashabaan48@gmail.com"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "{url}/api/{apiVersion}",
      "variables": {
        "url": {
          "default": "http://localhost:3000",
          "enum": [
            "http://localhost:3000",
            "https://ecommerece-api.onrender.com/"
          ]
        },
        "apiVersion": {
          "default": "v1"
        }
      }
    }
  ],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "description": "Get all products",
        "operationId": "getProducts",
        "respones": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/definations/Product"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "description": "Create a new product",
        "operationId": "createProduct",
        "tags": ["Products"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definations/Product"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definations/Product"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a product by id",
        "description": "Get a product by id",
        "operationId": "getProductById",
        "tags": ["Product"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definations/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a product by id",
        "description": "Update a product by id",
        "operationId": "updateProductById",
        "tags": ["update Product"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definations/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definations/Product"
                }
              }
            }
          }
        }
      },

      "delete": {
        "summary": "Delete a product by id",
        "description": "Delete a product by id",
        "operationId": "deleteProductById",
        "tags": ["delete Product"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definations/Product"
                }
              }
            }
          }
        }
      }
    },
...
    "/reviews": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get all reviews",
        "description": "Get all reviews",
        "operationId": "getReviews",
        "respones": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/definations/Review"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new review",
        "description": "Create a new review",
        "operationId": "createReview",
        "tags": ["Reviews"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definations/Review"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definations/Review"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{id}": {
      "get": {
        "summary": "Get a review by id",
        "description": "Get a review by id",
        "operationId": "getReviewById",
        "tags": ["Review"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/definations/Review"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a review by id",
        "description": "Update a review by id",
        "operationId": "updateReviewById",
        "tags": ["update Review"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definations/Review"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/definations/Review"
                }
              }
            }
          }
        }
      },

      "delete": {
        "summary": "Delete a review by id",
        "description": "Delete a review by id",
        "operationId": "deleteReviewById",
        "tags": ["delete Review"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/definations/Review"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}